+++ CHANGELOG.rst
@@ -5,6 +5,9 @@ Changelog
 Next
 ====
 
+- Improve Datasette detection (#399)
+- Fix generic JSON handling of null values (#399)
+
 Version 1.2.8 - 2023-10-21
 ==========================